Add configure menu option
