+++ tea/ui/__init__.py
@@ -3,6 +3,7 @@
 # tea/ui/__init__.py
 from .menus import menu_screen
 from .scans import discovery_scan_menu, full_scan_menu, schedule_scan_menu
+from .utils import config_screen, help_screen
 from .views import view_exposure
 from .welcome import show_welcome_screen
 
@@ -13,4 +14,6 @@ __all__ = [
     "view_exposure",
     "show_welcome_screen",
     "schedule_scan_menu",
+    "config_screen",
+    "help_screen",
 ]
