+++ tea/utils/startup_actions.py
@@ -4,7 +4,7 @@ import logging
 import os
 from pathlib import Path
 
-from dotenv import load_dotenv
+from dotenv import load_dotenv, set_key
 
 logger = logging.getLogger(__name__)
 
@@ -16,22 +16,25 @@ def startup_actions() -> None:
     This function needs to be called at the beginning of the program to ensure
     that all required initializations and configurations are done.
     """
+    # Set the environment path
     tea_root = Path(__file__).resolve().parent.parent.parent
-
+    dotenv_path = tea_root / ".env"
+    set_key(dotenv_path=dotenv_path, key_to_set="TEA_ROOT", value_to_set=str(tea_root))
+    
     # Load variables from the .env file into the environment
-    load_dotenv(dotenv_path=tea_root / ".env")
+    load_dotenv(dotenv_path=dotenv_path)
 
     # Clear existing handlers (useful in testing)
     logging.root.handlers.clear()
 
+    # Set or retrieve log path
     log_path = os.getenv("LOG_PATH")
     if not log_path:
         log_path = str(tea_root / ".tea.log")
-        os.environ["LOG_PATH"] = log_path
-    else:
-        log_path = Path(log_path)
-
+        set_key(dotenv_path=dotenv_path, key_to_set="LOG_PATH", value_to_set=log_path)
+    
     # Configure logging
+    log_path = Path(log_path)
     logging.basicConfig(
         level=logging.DEBUG,
         format="%(asctime)s [%(levelname)s] %(name)s: %(message)s",
@@ -39,13 +42,13 @@ def startup_actions() -> None:
     )
 
     # Check the DB-path and set to default if not set
-    db_path: str = os.getenv("EXPOSURE_DB_PATH")
+    db_path = os.getenv("EXPOSURE_DB_PATH")
     if not db_path:
         db_path = str(tea_root / ".exposure.sqlite")
-        os.environ["EXPOSURE_DB_PATH"] = db_path
+        set_key(dotenv_path=dotenv_path, key_to_set="EXPOSURE_DB_PATH", value_to_set=db_path)
 
     # Check the schedule path and set to default if not set
-    schedule_path: str = os.getenv("SCHEDULE_PATH")
+    schedule_path = os.getenv("SCHEDULE_PATH")
     if not schedule_path:
         schedule_path = str(tea_root / "schedule.json")
-        os.environ["SCHEDULE_PATH"] = schedule_path
+        set_key(dotenv_path=dotenv_path, key_to_set="SCHEDULE_PATH", value_to_set=schedule_path)
